requirements.txt
requests
pytest
python-dotenv
pytest-pyreport
